Refactor imports for @hedvig-ui/FadeIn
