+++ src/components/claims/claim-details/index.tsx
@@ -1,9 +1,8 @@
 import styled from '@emotion/styled'
-import { Card, CardsWrapper, MainHeadline } from '@hedvig-ui'
+import { Card, CardsWrapper, FadeIn, MainHeadline } from '@hedvig-ui'
 import { ClaimState, useClaimPageQuery } from 'api/generated/graphql'
 import { ClaimItems } from 'components/claims/claim-details/components/claim-items'
 import { ChatPane } from 'components/member/tabs/ChatPane'
-import { FadeIn } from 'hedvig-ui/animations/fade-in'
 import React, { useContext, useEffect } from 'react'
 import { Prompt, RouteComponentProps } from 'react-router'
 import { MemberHistoryContext } from 'utils/member-history'
