+++ src/components/claims/claims-list/components/ClaimListItem.tsx
@@ -1,11 +1,10 @@
 import styled from '@emotion/styled'
-import { Badge, Capitalized } from '@hedvig-ui'
+import { Badge, Capitalized, withFadeIn } from '@hedvig-ui'
 import { Claim, ClaimState } from 'api/generated/graphql'
 import { LinkRow } from 'components/shared'
 import { parseISO } from 'date-fns'
 import formatDate from 'date-fns/format'
 import isValidDate from 'date-fns/isValid'
-import { withFadeIn } from 'hedvig-ui/animations/fade-in'
 import React from 'react'
 import { useHistory } from 'react-router'
 import { Table, TableRowProps } from 'semantic-ui-react'
