+++ src/components/claims/index.tsx
@@ -1,9 +1,8 @@
-import { LoadingMessage, Spacing } from '@hedvig-ui'
+import { FadeIn, LoadingMessage, Spacing } from '@hedvig-ui'
 import { Claim } from 'api/generated/graphql'
 import { ListPage } from 'components/shared'
 import { Paginator } from 'components/shared/paginator/Paginator'
 import { useListClaims } from 'graphql/use-list-claims'
-import { FadeIn } from 'hedvig-ui/animations/fade-in'
 import React, { useEffect } from 'react'
 import { RouteComponentProps, useHistory } from 'react-router'
 import { Header } from 'semantic-ui-react'
