+++ src/components/dashboard/index.tsx
@@ -3,6 +3,7 @@ import {
   Capitalized,
   CasualList,
   CasualListItem,
+  FadeIn,
   MainHeadline,
   SecondLevelHeadline,
   Spacing,
@@ -13,7 +14,6 @@ import { changelog } from 'changelog'
 import { NumberMemberGroupsRadioButtons } from 'components/questions/number-member-groups-radio-buttons'
 import { differenceInCalendarDays, format } from 'date-fns'
 import { useDashboardNumbers } from 'graphql/use-dashboard-numbers'
-import { FadeIn } from 'hedvig-ui/animations/fade-in'
 import { Badge } from 'hedvig-ui/Badge/badge'
 import React from 'react'
 import { Link } from 'react-router-dom'
