+++ src/components/member/tabs/campaigns-tab/index.tsx
@@ -1,9 +1,13 @@
-import { LoadingMessage, MainHeadline, StandaloneMessage } from '@hedvig-ui'
+import {
+  FadeIn,
+  LoadingMessage,
+  MainHeadline,
+  StandaloneMessage,
+} from '@hedvig-ui'
 import { CampaignsInfo } from 'components/member/tabs/campaigns-tab/campaigns/CampaignsInfo'
 import { ReferralsInfo } from 'components/member/tabs/campaigns-tab/referrals/ReferralsInfo'
 import { RefreshButton } from 'components/member/tabs/shared/refresh-button'
 import { useGetReferralInformation } from 'graphql/use-get-referral-information'
-import { FadeIn } from 'hedvig-ui/animations/fade-in'
 import React from 'react'
 import { ArrowRepeat } from 'react-bootstrap-icons'
