+++ src/components/member/tabs/contracts-tab/index.tsx
@@ -1,10 +1,14 @@
-import { LoadingMessage, MainHeadline, StandaloneMessage } from '@hedvig-ui'
+import {
+  FadeIn,
+  LoadingMessage,
+  MainHeadline,
+  StandaloneMessage,
+} from '@hedvig-ui'
 import { Trial, useGetTrialsQuery } from 'api/generated/graphql'
 import { Contract } from 'components/member/tabs/contracts-tab/contract'
 import { TrialComponent } from 'components/member/tabs/contracts-tab/trial'
 import { RefreshButton } from 'components/member/tabs/shared/refresh-button'
 import { useContracts } from 'graphql/use-contracts'
-import { FadeIn } from 'hedvig-ui/animations/fade-in'
 import React from 'react'
 import { ArrowRepeat } from 'react-bootstrap-icons'
